+++ core/src/thing/labelForType.ts
@@ -1,0 +1,7 @@
+export function labelForType(typeUri: string) {
+  if (typeUri.includes("#")) {
+    return typeUri.substring(typeUri.lastIndexOf("#") + 1);
+  } else {
+    return typeUri.substring(typeUri.lastIndexOf("/") + 1);
+  }
+}